Improve provision
- Port some of the pending provisions to bash scripts
- Handle git diff in the `n` command
- Add new applications in the nix config (stripe, bitwarden)
- Better mappings handling in nvim
- General scripts improvements

+++ unix/scripts/copy_custom.sh
@@ -6,7 +6,7 @@ cd ~/development/environment
 
 BRANCH_NAME=$(git rev-parse --abbrev-ref HEAD)
 
-if [ "$BRANCH_NAME" != "custom" ] && [ "$(git status --porcelain | wc -l)" != "0" ]; then
+if [ "$BRANCH_NAME" != "custom" ] && [ "$(git status --porcelain | wc -l | sed 's| ||g')" != "0" ]; then
   echo "You have uncommitted changes, exiting..."
   exit 1
 fi
@@ -17,12 +17,22 @@ fi
 
 # If updating this, remember to update ./diff_custom.sh
 
-rsync -rhv --delete ./project/scripts/ custom/scripts/
+if [ -d project/scripts ]; then
+  rsync -rhv --delete ./project/scripts/ custom/scripts/
+fi
+if [ -d ~/development/nix-envs ]; then
+  rsync -rhv --delete ~/development/nix-envs/ custom/nix-envs/
+fi
+
 rsync -rhv --delete --exclude=.gitignore unix/scripts/bootstrap/ custom/bootstrap/
 rsync -rhv --delete src/custom.sh custom/custom.sh
 rsync -rhv --delete project/.vim-custom.lua custom/.vim-custom.lua
 rsync -rhv --delete project/vim-macros-custom custom/vim-macros-custom
 rsync -rhv --delete project/custom_create_vim_snippets.sh custom/custom_create_vim_snippets.sh
+
+if [ -n "$JUST_COMMIT" ]; then
+  exit 0
+fi
 
 git add -A .
 git commit -m "Update custom branch" || true
